+++ packages/rust-bindings/src/aux_functions.rs
@@ -3,7 +3,7 @@ use crypto_common::{types::TransactionTime, *};
 use dodis_yampolskiy_prf as prf;
 use pairing::bls12_381::{Bls12, G1};
 use serde_json::{from_str, Value as SerdeValue};
-use std::{cmp::max, collections::BTreeMap, convert::TryInto};
+use std::{collections::BTreeMap, convert::TryInto};
 type ExampleCurve = G1;
 use concordium_contracts_common::{from_bytes, schema, Cursor};
 use hex;
@@ -42,7 +42,7 @@ pub struct IdRequestInput {
     seed:           String,
     net:            String,
     identity_index: u32,
-    ar_threshold:   Option<u8>,
+    ar_threshold:   u8,
 }
 
 pub fn create_id_request_v1_aux(input: IdRequestInput) -> Result<String> {
@@ -69,25 +69,14 @@ pub fn create_id_request_v1_aux(input: IdRequestInput) -> Result<String> {
         wallet.get_blinding_randomness(input.identity_index)?;
 
     let num_of_ars = input.ars_infos.len();
-    let threshold = match input.ar_threshold {
-        Some(threshold) => {
-            ensure!(threshold > 0, "arThreshold must be at least 1.");
-            ensure!(
-                num_of_ars >= usize::from(threshold),
-                "Number of anonymity revokers in arsInfos should be at least arThreshold."
-            );
-            Threshold(threshold)
-        }
-        None => {
-            // arThreshold not specified, use `number of anonymity revokers` - 1 or 1 in the
-            // case of only a single anonymity revoker.
-            ensure!(
-                num_of_ars > 0,
-                "arsInfos should have at least 1 anonymity revoker."
-            );
-            Threshold(max((num_of_ars - 1).try_into().unwrap_or(255), 1))
-        }
-    };
+
+    ensure!(input.ar_threshold > 0, "arThreshold must be at least 1.");
+    ensure!(
+        num_of_ars >= usize::from(input.ar_threshold),
+        "Number of anonymity revokers in arsInfos should be at least arThreshold."
+    );
+
+    let threshold = Threshold(input.ar_threshold);
 
     let chi = CredentialHolderInfo::<ArCurve> { id_cred };
 
@@ -96,7 +85,6 @@ pub fn create_id_request_v1_aux(input: IdRequestInput) -> Result<String> {
         prf_key,
     };
 
-    // Choice of anonymity revokers, all of them in this implementation.
     let context = IpContext::new(&input.ip_info, &input.ars_infos, &input.global_context);
 
     let id_use_data = IdObjectUseData {